requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

omegaconf==2.3.0
jaxtyping
typeguard
nerfacc==0.3.3
pyrender
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
diffusers[torch]==0.19.3
transformers==4.28.1
accelerate
opencv-python
tensorboard
matplotlib
imageio>=2.28.0
imageio[ffmpeg]
git+https://github.com/NVlabs/nvdiffrast.git
libigl
xatlas
trimesh[easy]
networkx
pysdf
PyMCubes==0.1.2
wandb
gradio
git+https://github.com/ashawkey/envlight.git
torchmetrics

# deepfloyd
xformers==0.0.22.post7+cu118
bitsandbytes==0.38.1
sentencepiece
safetensors
huggingface_hub

# for zero123
einops
kornia==0.7.0
taming-transformers-rom1504
git+https://github.com/openai/CLIP.git

#controlnet
controlnet_aux

#vis
torch_efficient_distloss
mediapy
plyfile
rembg
git+https://github.com/ashawkey/kiuikit
scikit-learn
scipy
viser
ninja
easydict
albumentations==0.5.2
webdataset
git+https://github.com/luca-medeiros/lang-segment-anything.git


#wonder3d
nerfacc==0.3.3
pyrender==0.1.45
Pillow==9.5.0


# gaussian splatting
gaussiansplatting/submodules/diff-gaussian-rasterization
gaussiansplatting/submodules/simple-knn